Proceed ruler's figure part zoom.

+++ edu.xmu.nmr.dataAnalysis.diagram/src/edu/xmu/nmr/dataanalysis/diagram/layouts/CoordinateTf.java
@@ -8,103 +8,104 @@ import org.eclipse.draw2d.geometry.Point;
  * </p>
  * 
  * @author software
- *
+ *         
  */
 public class CoordinateTf {
-
-	/**
-	 * x轴比例
-	 */
-	private float xScale;
-	/**
-	 * y轴比例
-	 */
-	private float yScale;
-	/**
-	 * x轴偏移
-	 */
-	private float xOffset;
-	/**
-	 * y轴偏移
-	 */
-	private float yOffset;
-
-	public CoordinateTf(float xScale, float yScale, float xOffset, float yOffset) {
-		this.xScale = xScale;
-		this.yScale = yScale;
-		this.xOffset = xOffset;
-		this.yOffset = yOffset;
-	}
-
-	public CoordinateTf() {
-		this(1.0f, 1.0f, 0.0f, 0.0f);
-	}
-
-	/**
-	 * 将指定的数据点的x值转换到相应的坐标系下
-	 * 
-	 * @param sourX
-	 *            源数据的X值
-	 * @return int 指定坐标系下X
-	 */
-	public int transfromX(float sourX) {
-		return (int) (sourX * xScale + xOffset);
-	}
-
-	/**
-	 * 将指定的数据点的y值转换到相应的坐标系下
-	 * 
-	 * @param sourY
-	 *            源数据的Y值
-	 * @return int 指定坐标系下 Y
-	 */
-	public int transformY(float sourY) {
-		return (int) (sourY * yScale + yOffset);
-	}
-
-	public Point transformXY(float sourX, float sourY) {
-		Point p = new Point();
-		p.x = transfromX(sourX);
-		p.y = transformY(sourY);
-		return p;
-	}
-
-	public Point tranformPoint(Point sourPoint) {
-		Point p = new Point();
-		p.x = transfromX(sourPoint.x);
-		p.y = transformY(sourPoint.y);
-		return p;
-	}
-
-	public float getxScale() {
-		return xScale;
-	}
-
-	public void setxScale(float xScale) {
-		this.xScale = xScale;
-	}
-
-	public float getyScale() {
-		return yScale;
-	}
-
-	public void setyScale(float yScale) {
-		this.yScale = yScale;
-	}
-
-	public float getxOffset() {
-		return xOffset;
-	}
-
-	public void setxOffset(float xOffset) {
-		this.xOffset = xOffset;
-	}
-
-	public float getyOffset() {
-		return yOffset;
-	}
-
-	public void setyOffset(float yOffset) {
-		this.yOffset = yOffset;
-	}
+    
+    /**
+     * x轴比例
+     */
+    private double xScale;
+    /**
+     * y轴比例
+     */
+    private double yScale;
+    /**
+     * x轴偏移
+     */
+    private float xOffset;
+    /**
+     * y轴偏移
+     */
+    private float yOffset;
+    
+    public CoordinateTf(float xScale, float yScale, float xOffset,
+            float yOffset) {
+        this.xScale = xScale;
+        this.yScale = yScale;
+        this.xOffset = xOffset;
+        this.yOffset = yOffset;
+    }
+    
+    public CoordinateTf() {
+        this(1.0f, 1.0f, 0.0f, 0.0f);
+    }
+    
+    /**
+     * 将指定的数据点的x值转换到相应的坐标系下
+     * 
+     * @param sourX
+     *            源数据的X值
+     * @return int 指定坐标系下X
+     */
+    public int transfromX(float sourX) {
+        return (int) (sourX * xScale + xOffset);
+    }
+    
+    /**
+     * 将指定的数据点的y值转换到相应的坐标系下
+     * 
+     * @param sourY
+     *            源数据的Y值
+     * @return int 指定坐标系下 Y
+     */
+    public int transformY(float sourY) {
+        return (int) (sourY * yScale + yOffset);
+    }
+    
+    public Point transformXY(float sourX, float sourY) {
+        Point p = new Point();
+        p.x = transfromX(sourX);
+        p.y = transformY(sourY);
+        return p;
+    }
+    
+    public Point tranformPoint(Point sourPoint) {
+        Point p = new Point();
+        p.x = transfromX(sourPoint.x);
+        p.y = transformY(sourPoint.y);
+        return p;
+    }
+    
+    public double getxScale() {
+        return xScale;
+    }
+    
+    public void setxScale(double xScale) {
+        this.xScale = xScale;
+    }
+    
+    public double getyScale() {
+        return yScale;
+    }
+    
+    public void setyScale(double yScale) {
+        this.yScale = yScale;
+    }
+    
+    public float getxOffset() {
+        return xOffset;
+    }
+    
+    public void setxOffset(float xOffset) {
+        this.xOffset = xOffset;
+    }
+    
+    public float getyOffset() {
+        return yOffset;
+    }
+    
+    public void setyOffset(float yOffset) {
+        this.yOffset = yOffset;
+    }
 }
